% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardMethods.R
\docType{class}
\name{figOutput-class}
\alias{figOutput-class}
\title{S4 Class for displaying figures}
\description{
S4 Class to hold the figures output from the getFigure function.
}
\section{Slots}{

\describe{
\item{\code{figures}}{A list of lists containing figures for each specified sub_model and y_model (when
applicable).}
}}

