% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardMethods.R
\docType{class}
\name{FSOutput-class}
\alias{FSOutput-class}
\title{S4 Class for estimated factor scores and their standard errors.}
\description{
S4 Class for the output structure for the \code{getIndFS()} function.
}
\section{Slots}{

\describe{
\item{\code{scores_est}}{A matrix of estimated factor scores.}

\item{\code{scores_se}}{A matrix of standard errors of estimated factor scores.}
}}

