% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGCM.mxModel_helper.R
\name{getLGCM.mxModel}
\alias{getLGCM.mxModel}
\title{Construct An Object of mxModel for Latent Growth Curve Model with Time-invariant Covariates (If Any) To Be
Evaluated}
\usage{
getLGCM.mxModel(
  dat,
  t_var,
  y_var,
  curveFun,
  intrinsic,
  records,
  growth_TIC,
  starts
)
}
\arguments{
\item{dat}{A wide-format data frame, with each row corresponding to a unique ID. It contains the observed variables with
repeated measurements and occasions, and time-invariant covariates (TICs) if any. It takes the value passed from \code{getLGCM()}.}

\item{t_var}{A string specifying the prefix of the column names corresponding to the time variable at each study wave.
It takes the value passed from \code{getLGCM()}.}

\item{y_var}{A string specifying the prefix of the column names corresponding to the outcome variable at each study wave.
It takes the value passed from \code{getLGCM()}.}

\item{curveFun}{A string specifying the functional form of the growth curve. Supported options for latent growth curve
models are: \code{"linear"} (or \code{"LIN"}), \code{"quadratic"} (or \code{"QUAD"}), \code{"negative exponential"}
(or \code{"EXP"}), \code{"Jenss-Bayley"} (or \code{"JB"}), and \code{"bilinear spline"} (or \code{"BLS"}).
It takes the value passed from \code{getLGCM()}.}

\item{intrinsic}{A logical flag indicating whether to build an intrinsically nonlinear longitudinal model. It takes the
value passed from \code{getLGCM()}.}

\item{records}{A numeric vector specifying indices of the study waves. It takes the value passed from \code{getLGCM()}.}

\item{growth_TIC}{A string or character vector specifying the column name(s) of time-invariant covariate(s) contributing to the
variability of growth factors if any. It takes the value passed from \code{getLGCM()}.}

\item{starts}{A list containing initial values for the free parameters, either takes the value passed from \code{getLGCM()} or
derived by the helper function \code{getUNI.initial()}.}
}
\value{
A pre-optimized mxModel for a Latent Growth Curve Model.
}
\description{
his function builds up an object of mxModel for a latent growth curve model with user-specified
functional form (including whether intrinsically nonlinear) with time-invariant covariates (if any).
}
\keyword{internal}
