% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGCM.output_helper.R
\name{getLGCM.output}
\alias{getLGCM.output}
\title{Extract Point Estimates and Standard Errors of Latent Growth Curve Model with Time-invariant Covariates (If Any)}
\usage{
getLGCM.output(model, growth_TIC, names)
}
\arguments{
\item{model}{An object representing a fitted latent growth curve model.}

\item{growth_TIC}{A string or character vector specifying the column name(s) of time-invariant covariate(s) contributing to the
variability of growth factors if any. It takes the value passed from \code{getLGCM()}.}

\item{names}{A character vector specifying parameter names. It takes the value passed from \code{getLGCM()}.}
}
\value{
A data frame containing the point estimates and standard errors for parameters of a latent growth curve
model with time-invariant covariates (if any).
}
\description{
This function computes and returns a data frame containing point estimates and standard errors for the parameters
of a latent growth curve model with time-invariant covariates (if any)
}
\keyword{internal}
