% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMGroup.R
\name{getMGroup}
\alias{getMGroup}
\title{Fit a Longitudinal Multiple Group Model}
\usage{
getMGroup(
  dat,
  grp_var,
  sub_Model,
  t_var,
  records,
  y_var,
  curveFun,
  intrinsic = NULL,
  y_model = NULL,
  m_var = NULL,
  x_type = NULL,
  x_var = NULL,
  TVC = NULL,
  decompose = NULL,
  growth_TIC = NULL,
  starts = NULL,
  res_scale = NULL,
  res_cor = NULL,
  tries = NULL,
  OKStatus = 0,
  jitterD = "runif",
  loc = 1,
  scale = 0.25,
  paramOut = FALSE,
  names = NULL
)
}
\arguments{
\item{dat}{A wide-format data frame, with each row corresponding to a unique ID. It contains the observed variables with
repeated measurements and occasions for each longitudinal process, time-invariant covariates (TICs) if any, and a variable
that indicates manifested group.}

\item{grp_var}{A string specifying the column that indicates manifested classes.}

\item{sub_Model}{A string that specifies the sub-model for manifested classes. Supported sub-models include \code{"LGCM"}
(for latent growth curve models), \code{"LCSM"} (for latent change score models), \code{"TVC"} (for latent growth curve
models or latent change score models with a time-varying covariate), \code{"MGM"} (for multivariate latent growth curve
models or latent change score models), and \code{"MED"} (for longitudinal mediation models).}

\item{t_var}{A string specifying the prefix of the column names corresponding to the time variable for each study wave.
This applies when \code{sub_Model} is \code{"LGCM"}, \code{"LCSM"} or \code{"TVC"}. For \code{sub_Model} being \code{"MGM"}
or \code{"MED"}, \code{t_var} should be a string vector where each element corresponds to the time variable prefix for each
respective longitudinal process.}

\item{records}{A numeric vector denoting the indices of the observed study waves. This applies when \code{sub_Model} is
\code{"LGCM"}, \code{"LCSM"} or \code{"TVC"}. For \code{sub_Model} being \code{"MGM"} or \code{"MED"}, \code{records} should
be a list of numeric vectors, where each vector provides the indices of the observed study waves for each longitudinal process.}

\item{y_var}{A string defining the prefix of the column names corresponding to the outcome variable for each study wave. This
is applicable when \code{sub_Model} is not \code{"MGM"}. For \code{sub_Model} being \code{"MGM"}, \code{y_var} should be a
string vector where each element corresponds to the prefix of the column names for each outcome variable across the study waves.}

\item{curveFun}{A string specifying the functional forms of the growth curve(s). Supported options for \code{y_model = "LGCM"}
include: \code{"linear"} (or \code{"LIN"}), \code{"quadratic"} (or \code{"QUAD"}), \code{"negative exponential"} (or \code{"EXP"}),
\code{"Jenss-Bayley"} (or \code{"JB"}), and \code{"bilinear spline"} (or \code{"BLS"}). Supported options for \code{y_model =
"LCSM"} include: \code{"nonparametric"} (or \code{"NonP"}), \code{"quadratic"} (or \code{"QUAD"}), \code{"negative exponential"}
(or \code{"EXP"}), and \code{"Jenss-Bayley"} (or \code{"JB"}).}

\item{intrinsic}{A logical flag indicating whether to build an intrinsically nonlinear longitudinal model. By default, this is
\code{NULL}, as it is unnecessary when \code{sub_Model} is \code{"MED"}.}

\item{y_model}{A string that specifies how to fit longitudinal outcomes. Supported values are \code{"LGCM"} and \code{"LCSM"}.
By default, this is \code{NULL} as this argument is only required when \code{sub_Model} is \code{"TVC"} or \code{"MGM"}.}

\item{m_var}{A string that specifies the prefix of the column names corresponding to the mediator variable at each study wave.
By default, this is \code{NULL} as this argument is only required when \code{sub_Model} is \code{"MED"}.}

\item{x_type}{A string indicating the type of predictor variable used in the model. Supported values are \code{"baseline"}
and \code{"longitudinal"}. By default, this is \code{NULL} as this argument is only required when \code{sub_Model} is \code{"MED"}.}

\item{x_var}{A string specifying the baseline predictor if \code{x_type = "baseline"}, or the prefix of the column names
corresponding to the predictor variable at each study wave if \code{x_type = "longitudinal"}. By default, this is \code{NULL} as
this argument is only required when \code{sub_Model} is \code{"MED"}.}

\item{TVC}{A string that specifies the prefix of the column names corresponding to the time-varying covariate at each time
point. By default, this is \code{NULL} as this argument is only required when \code{sub_Model} is \code{"TVC"}.}

\item{decompose}{An integer specifying the decomposition option for temporal states. Supported values include \code{0} (no
decomposition), \code{1} (decomposition with interval-specific slopes as temporal states), \code{2} (decomposition with interval-
specific changes as temporal states), and \code{3} (decomposition with change-from-baseline as temporal states). By default, this
is \code{NULL} as this argument is only required when \code{sub_Model} is \code{"TVC"}.}

\item{growth_TIC}{A string or character vector of column names of time-invariant covariate(s) accounting for the variability
of growth factors if any. Default is \code{NULL}, indicating no growth TICs present in the model.}

\item{starts}{A list containing initial values for the parameters. Default is \code{NULL}, indicating no user-specified
initial values.}

\item{res_scale}{An optional list where each element is a (vector of) numeric scaling factor(s) for residual variance to
calculate the corresponding initial value for a latent class, between \code{0} and \code{1} exclusive. Default is \code{NULL},
in which case data-driven residual variance estimation is used. If data-driven estimation fails, a heuristic of \code{0.1} is
applied as fallback.}

\item{res_cor}{An optional list where each element is a (vector of) numeric initial value(s) for residual correlation in each
class. Applicable when the sub_Model is \code{"TVC"} (when \code{decompose != 0}), \code{"MGM"}, or \code{"MED"}. Default is
\code{NULL}, in which case data-driven residual correlation estimation is used. If data-driven estimation fails, a heuristic
of \code{0.3} is applied as fallback.}

\item{tries}{An integer specifying the number of additional optimization attempts. Default is \code{NULL}.}

\item{OKStatus}{An integer (vector) specifying acceptable status codes for convergence. Default is \code{0}.}

\item{jitterD}{A string specifying the distribution for jitter. Supported values are: \code{"runif"} (uniform
distribution), \code{"rnorm"} (normal distribution), and \code{"rcauchy"} (Cauchy distribution). Default is \code{"runif"}.}

\item{loc}{A numeric value representing the location parameter of the jitter distribution. Default is \code{1}.}

\item{scale}{A numeric value representing the scale parameter of the jitter distribution. Default is \code{0.25}.}

\item{paramOut}{A logical flag indicating whether to output the parameter estimates and standard errors. Default is \code{FALSE}.}

\item{names}{A character vector specifying parameter names. Default is \code{NULL}, in which case
meaningful names are automatically generated based on the model configuration.}
}
\value{
An object of class \code{myMxOutput}. Depending on the \code{paramOut} argument, the object may contain the following slots:
\itemize{
  \item \code{mxOutput}: This slot contains the fitted longitudinal multiple group model. A summary of this model can be obtained using
  the \code{ModelSummary()} function.
  \item \code{Estimates} (optional): If \code{paramOut = TRUE}, a data frame with parameter estimates and standard errors. The content
  of this slot can be printed using the \code{printTable()} method for S4 objects.
}
}
\description{
This function fits a longitudinal multiple group model based on the specified sub-model. Supported submodels
include:
\itemize{
  \item Latent growth curve models,
  \item Latent change score models,
  \item Latent growth curve models or latent change score models with a time-varying covariate,
  \item Multivariate latent growth curve models or multivariate latent change score models,
  \item Longitudinal mediation models.
} For the first three submodels, time-invariant covariates are allowed.
}
\examples{
mxOption(model = NULL, key = "Default optimizer", "CSOLNP", reset = FALSE)
data("RMS_dat")
# Re-baseline the data so that the estimated initial status is for the starting point of the study
RMS_dat0 <- RMS_dat
baseT <- RMS_dat0$T1
RMS_dat0$T1 <- RMS_dat0$T1 - baseT
RMS_dat0$T2 <- RMS_dat0$T2 - baseT
RMS_dat0$T3 <- RMS_dat0$T3 - baseT
RMS_dat0$T4 <- RMS_dat0$T4 - baseT
RMS_dat0$T5 <- RMS_dat0$T5 - baseT
RMS_dat0$T6 <- RMS_dat0$T6 - baseT
RMS_dat0$T7 <- RMS_dat0$T7 - baseT
RMS_dat0$T8 <- RMS_dat0$T8 - baseT
RMS_dat0$T9 <- RMS_dat0$T9 - baseT
RMS_dat0$ex1 <- scale(RMS_dat0$Approach_to_Learning)
RMS_dat0$ex2 <- scale(RMS_dat0$Attention_focus)
\donttest{
# Fit longitudinal multiple group model of bilinear spline functional form with fixed knot
MGroup_BLS_LGCM.TIC_f <-  getMGroup(
  dat = RMS_dat0, grp_var = "SEX", sub_Model = "LGCM", y_var = "M", t_var = "T",
  records = 1:9, curveFun = "BLS", intrinsic = FALSE
)

# Fit longitudinal multiple group model of bilinear spline functional form with random knot
set.seed(20191029)
MGroup_BLS_LGCM.TIC_f <-  getMGroup(
  dat = RMS_dat0, grp_var = "SEX", sub_Model = "LGCM", y_var = "M", t_var = "T",
  records = 1:9, curveFun = "BLS", intrinsic = TRUE,
  growth_TIC = c("ex1", "ex2"), tries = 10, paramOut = TRUE
)
printTable(MGroup_BLS_LGCM.TIC_f)
}

}
\references{
\itemize{
  \item {Liu, J., & Perera, R. A. (2023). Estimating Rate of Change for Nonlinear Trajectories in the Framework of Individual Measurement
  Occasions: A New Perspective on Growth Curves. Behavior Research Methods, 56(3), 1349-1375. \doi{10.3758/s13428-023-02097-2}}
  }
}
\seealso{
\code{\link{getLGCM}}, \code{\link{getLCSM}}, \code{\link{getMIX}}, \code{\link{getFigure}}
}
