% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardMethods.R
\docType{class}
\name{KappaOutput-class}
\alias{KappaOutput-class}
\title{S4 Class for kappa statistic with confidence interval and judgment.}
\description{
S4 Class for the output structure for the \code{getLatentKappa()} function.
}
\section{Slots}{

\describe{
\item{\code{kappa_value}}{A character vector for the kappa statistic with $95\%$ CI.}

\item{\code{judgment}}{A character vector for the judgement for agreement.}
}}

