% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardMethods.R
\name{ModelSummary,myMxOutput-method}
\alias{ModelSummary,myMxOutput-method}
\title{S4 Method for summarizing an optimized MxModel.}
\usage{
\S4method{ModelSummary}{myMxOutput}(object)
}
\arguments{
\item{object}{An object of class "myMxOutput".}
}
\description{
Method for printing model summary of MxModel object.
}
