% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardMethods.R
\name{ModelSummary}
\alias{ModelSummary}
\title{S4 Generic for summarizing an optimized MxModel.}
\usage{
ModelSummary(object)
}
\arguments{
\item{object}{An object of the appropriate class.}
}
\value{
Invisibly returns the summary of the MxModel object. Called primarily for its
side effect of printing the model summary.
}
\description{
Generic function for printing model summary of MxModel object.
}
\examples{
\donttest{
mxOption(model = NULL, key = "Default optimizer", "CSOLNP", reset = FALSE)
data("RMS_dat")
RMS_dat0 <- RMS_dat
baseT <- RMS_dat0$T1
RMS_dat0$T1 <- RMS_dat0$T1 - baseT
RMS_dat0$T2 <- RMS_dat0$T2 - baseT
RMS_dat0$T3 <- RMS_dat0$T3 - baseT
RMS_dat0$T4 <- RMS_dat0$T4 - baseT
RMS_dat0$T5 <- RMS_dat0$T5 - baseT
RMS_dat0$T6 <- RMS_dat0$T6 - baseT
RMS_dat0$T7 <- RMS_dat0$T7 - baseT
RMS_dat0$T8 <- RMS_dat0$T8 - baseT
RMS_dat0$T9 <- RMS_dat0$T9 - baseT
BLS_LGCM <- getLGCM(
  dat = RMS_dat0, t_var = "T", y_var = "M", curveFun = "BLS", intrinsic = FALSE,
  records = 1:9
)
ModelSummary(BLS_LGCM)
}
}
