% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardMethods.R
\docType{class}
\name{myMxOutput-class}
\alias{myMxOutput-class}
\title{Standard Methods (S4) for the package}
\description{
S4 Class for the output structure for estimate functions.
}
\section{Slots}{

\describe{
\item{\code{mxOutput}}{An object of class "MxModel".}

\item{\code{Estimates}}{A data frame of estimates.}
}}

