% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardMethods.R
\docType{class}
\name{postOutput-class}
\alias{postOutput-class}
\title{S4 Class for posterior probabilities, membership, entropy, and accuracy (when applicable)}
\description{
S4 Class for the output structure for the \code{getPosterior()} function.
}
\section{Slots}{

\describe{
\item{\code{prob}}{A matrix of posterior probabilities.}

\item{\code{membership}}{A numeric vector for membership.}

\item{\code{entropy}}{A numeric value for entropy.}

\item{\code{accuracy}}{A numeric value for accuracy.}
}}

