% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardMethods.R
\name{printTable,FSOutput-method}
\alias{printTable,FSOutput-method}
\title{S4 Method for printing estimated factor scores and their standard errors}
\usage{
\S4method{printTable}{FSOutput}(object)
}
\arguments{
\item{object}{An object of class "FSOutput".}
}
\value{
Called for its side effect of printing factor scores and their standard errors.
Returns the object invisibly.
}
\description{
Method for printing estimated factor scores and their standard errors.
}
