% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardMethods.R
\name{printTable,KappaOutput-method}
\alias{printTable,KappaOutput-method}
\title{S4 Method for printing kappa statistic with $95\%$ CI and judgement for agreement.}
\usage{
\S4method{printTable}{KappaOutput}(object)
}
\arguments{
\item{object}{An object of class "KappaOutput".}
}
\description{
Method for printing kappa statistic with $95\%$ CI and judgement for agreement.
}
