% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardMethods.R
\name{printTable,myMxOutput-method}
\alias{printTable,myMxOutput-method}
\title{S4 Method for printing point estimates with standard errors}
\usage{
\S4method{printTable}{myMxOutput}(object)
}
\arguments{
\item{object}{An object of class "myMxOutput".}
}
\description{
Method for printing point estimates and standard errors.
}
