% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardMethods.R
\name{printTable,postOutput-method}
\alias{printTable,postOutput-method}
\title{S4 Method for printing posterior probabilities, membership, entropy, and accuracy.}
\usage{
\S4method{printTable}{postOutput}(object)
}
\arguments{
\item{object}{An object of class "postOutput".}
}
\description{
Method for printing posterior probabilities, membership, entropy, and accuracy.
}
