% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardMethods.R
\name{printTable,StatsOutput-method}
\alias{printTable,StatsOutput-method}
\title{S4 Method for printing p values and confidence intervals (when applicable)}
\usage{
\S4method{printTable}{StatsOutput}(object)
}
\arguments{
\item{object}{An object of class "StatsOutput".}
}
\description{
Method for printing p values and confidence intervals.
}
