% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardMethods.R
\name{show,figOutput-method}
\alias{show,figOutput-method}
\title{S4 Method for displaying figures.}
\usage{
\S4method{show}{figOutput}(object)
}
\arguments{
\item{object}{An object of class "figOutput".}
}
\description{
Method to display a summary of the figOutput object when printed.
}
