% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardMethods.R
\docType{class}
\name{StatsOutput-class}
\alias{StatsOutput-class}
\title{S4 Class for p values and confidence intervals (when specified).}
\description{
S4 Class for the output structure for the \code{getEstimateStats()} function.
}
\section{Slots}{

\describe{
\item{\code{wald}}{A data frame for p values and Wald confidence intervals (when specified).}

\item{\code{likelihood}}{A data frame for Likelihood confidence intervals (when specified).}

\item{\code{bootstrap}}{A data frame for Bootstrap confidence intervals (when specified).}
}}

