% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardMethods.R
\name{ModelSummary}
\alias{ModelSummary}
\title{S4 Generic for summarizing an optimized MxModel.}
\usage{
ModelSummary(object)
}
\arguments{
\item{object}{An object of the appropriate class.}
}
\description{
Generic function for printing model summary of MxModel object.
}
