% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardMethods.R
\name{printTable}
\alias{printTable}
\title{S4 Generic for displaying output in a table format.}
\usage{
printTable(object)
}
\arguments{
\item{object}{An object of the appropriate class.}
}
\description{
Generic function for printing output that are tables.
}
